move scaled activations inside custom op
